+++ src/eps/impls/krylov/krylovschur/krylovschur.h
@@ -26,6 +26,8 @@ SLEPC_INTERN PetscErrorCode EPSSolve_KrylovSchur_BSE_Gruning(EPS);
 SLEPC_INTERN PetscErrorCode EPSSolve_KrylovSchur_BSE_ProjectedBSE(EPS);
 SLEPC_INTERN PetscErrorCode EPSSetUp_KrylovSchur_Hamilt(EPS);
 SLEPC_INTERN PetscErrorCode EPSSolve_KrylovSchur_Hamilt(EPS);
+SLEPC_INTERN PetscErrorCode EPSSetUp_KrylovSchur_LREP(EPS);
+SLEPC_INTERN PetscErrorCode EPSSolve_KrylovSchur_LREP_Teng(EPS);
 SLEPC_INTERN PetscErrorCode EPSGetArbitraryValues(EPS,PetscScalar*,PetscScalar*);
 SLEPC_INTERN PetscErrorCode EPSKrylovSchurGetChildEPS(EPS,EPS*);