+++ src/sys/mat/matstruct.c
@@ -41,7 +41,7 @@
 
    Level: intermediate
 
-.seealso: [](sec:structured), `MatCreateNest()`, `EPSSetOperators()`, `EPSSetProblemType()`, `MatCreateHamiltonian()`
+.seealso: [](sec:structured), `MatCreateNest()`, `EPSSetOperators()`, `EPSSetProblemType()`, `MatCreateHamiltonian()`, `MatCreateLREP()`
 @*/
 PetscErrorCode MatCreateBSE(Mat R,Mat C,Mat *H)
 {
@@ -111,7 +111,7 @@ PetscErrorCode MatCreateBSE(Mat R,Mat C,Mat *H)
 
    Level: intermediate
 
-.seealso: [](sec:structured), `MatCreateNest()`, `EPSSetOperators()`, `EPSSetProblemType()`, `MatCreateBSE()`
+.seealso: [](sec:structured), `MatCreateNest()`, `EPSSetOperators()`, `EPSSetProblemType()`, `MatCreateBSE()`, `MatCreateLREP()`
 @*/
 PetscErrorCode MatCreateHamiltonian(Mat A,Mat B,Mat C,Mat *H)
 {
@@ -148,5 +148,111 @@ PetscErrorCode MatCreateHamiltonian(Mat A,Mat B,Mat C,Mat *H)
   PetscCall(PetscNew(&mctx));
   mctx->cookie = SLEPC_MAT_STRUCT_HAMILT;
   PetscCall(PetscObjectContainerCompose((PetscObject)*H,"SlepcMatStruct",mctx,PetscCtxDestroyDefault));
+  PetscFunctionReturn(PETSC_SUCCESS);
+}
+
+/*@
+   MatCreateLREP - Create a matrix that can be used to define a structured Linear
+   Response eigenvalue problem.
+
+   Collective
+
+   Input Parameters:
++  AK  - matrix for the diagonal block ($A$) or the top block ($K$)
+.  BM  - matrix for the off-diagonal block ($B$) or the bottom block ($M$)
+-  red - whether the reduced form should be built
+
+   Output Parameter:
+.  H  - the resulting matrix
+
+   Notes:
+   The resulting matrix has the block form $$H = \begin{bmatrix} A & B \\ -B & -A \end{bmatrix},$$
+   where both $A$ and $B$ are assumed to be real symmetric. An alternative form,
+   called reduced form (`red`=`PETSC_TRUE`) is $$H = \begin{bmatrix} 0 & K \\ M & 0 \end{bmatrix},$$
+   where both $K$ and $M$ are also assumed to be real symmetric. Note that this function
+   does not check these properties, so if the matrices provided by the user do not
+   satisfy them, then the solver will not behave as expected.
+
+   The reduced form can be obtained from the original formulation by setting $K=A-B$
+   and $M=A+B$. However, $K$ and $M$ need not have such relation, in which case the
+   computed eigenvalues will be $\pm\lambda_i$, where $\lambda_i^2$ are the eigenvalues
+   of the _product eigenvalue problem_ $KM$ (or $MK$).
+
+   The obtained matrix can be used as an input matrix to `EPS` eigensolvers via
+   `EPSSetOperators()` for the case that the problem type is `EPS_LREP`. Note that the
+   user cannot just build a matrix with the required structure, it must be done via
+   this function.
+
+   In the current implementation, `H` is a `MATNEST` matrix, where `A` and `B` form the
+   top block row, while the bottom block row is composed of shell matrices scaled by -1.
+   Similarly, in the reduced formulation, the `MATNEST` has two zero blocks.
+
+   Internally, the eigensolver will work with the reduced form, and it will carry out
+   the conversion transparently if the first form is provided by the user.
+
+   Level: intermediate
+
+.seealso: [](sec:structured), `MatCreateNest()`, `EPSSetOperators()`, `EPSSetProblemType()`, `MatCreateBSE()`, `MatCreateHamiltonian()`
+@*/
+PetscErrorCode MatCreateLREP(Mat AK,Mat BM,PetscBool red,Mat *H)
+{
+  PetscInt       Ma,Mb,Na,Nb,ma,mb,na,nb;
+  SlepcMatStruct mbtx;
+
+  PetscFunctionBegin;
+  PetscValidHeaderSpecific(AK,MAT_CLASSID,1);
+  PetscValidHeaderSpecific(BM,MAT_CLASSID,2);
+  PetscCheckSameTypeAndComm(AK,1,BM,2);
+  PetscValidLogicalCollectiveBool(AK,red,3);
+  PetscAssertPointer(H,4);
+
+  /* check sizes */
+  PetscCall(MatGetSize(AK,&Ma,&Na));
+  PetscCall(MatGetLocalSize(AK,&ma,&na));
+  PetscCall(MatGetSize(BM,&Mb,&Nb));
+  PetscCall(MatGetLocalSize(BM,&mb,&nb));
+  PetscCheck(Mb==Ma && mb==ma,PetscObjectComm((PetscObject)AK),PETSC_ERR_ARG_INCOMP,"Incompatible row dimensions");
+  PetscCheck(Nb==Na && nb==na,PetscObjectComm((PetscObject)AK),PETSC_ERR_ARG_INCOMP,"Incompatible column dimensions");
+
+  if (red) {  /* reduced form */
+
+    Mat block[4] = { NULL, AK, BM, NULL };
+
+    /* create nest matrix */
+    PetscCall(MatCreateNest(PetscObjectComm((PetscObject)AK),2,NULL,2,NULL,block,H));
+
+  } else {  /* original LREP form */
+
+    Mat block[4] = { AK, BM, NULL, NULL };
+    const PetscScalar scal[] = { -1.0 };
+
+    /* bottom block row, using MATCOMPOSITE */
+    PetscCall(MatCreate(PetscObjectComm((PetscObject)AK),&block[2]));
+    PetscCall(MatSetSizes(block[2],mb,nb,Mb,Nb));
+    PetscCall(MatSetType(block[2],MATCOMPOSITE));
+    PetscCall(MatCompositeAddMat(block[2],BM));
+    PetscCall(MatAssemblyBegin(block[2],MAT_FINAL_ASSEMBLY));
+    PetscCall(MatAssemblyEnd(block[2],MAT_FINAL_ASSEMBLY));
+    PetscCall(MatCompositeSetScalings(block[2],scal));
+
+    PetscCall(MatCreate(PetscObjectComm((PetscObject)AK),&block[3]));
+    PetscCall(MatSetSizes(block[3],ma,na,Ma,Na));
+    PetscCall(MatSetType(block[3],MATCOMPOSITE));
+    PetscCall(MatCompositeAddMat(block[3],AK));
+    PetscCall(MatAssemblyBegin(block[3],MAT_FINAL_ASSEMBLY));
+    PetscCall(MatAssemblyEnd(block[3],MAT_FINAL_ASSEMBLY));
+    PetscCall(MatCompositeSetScalings(block[3],scal));
+
+    /* create nest matrix */
+    PetscCall(MatCreateNest(PetscObjectComm((PetscObject)AK),2,NULL,2,NULL,block,H));
+    PetscCall(MatDestroy(&block[2]));
+    PetscCall(MatDestroy(&block[3]));
+
+  }
+
+  /* compose context */
+  PetscCall(PetscNew(&mbtx));
+  mbtx->cookie = SLEPC_MAT_STRUCT_LREP;
+  PetscCall(PetscObjectContainerCompose((PetscObject)*H,"SlepcMatStruct",mbtx,PetscCtxDestroyDefault));
   PetscFunctionReturn(PETSC_SUCCESS);
 }